Move utilities from ros pkg to common pkg and add missing dependencies cmake config
