+++ rct_common/CMakeLists.txt
@@ -3,6 +3,17 @@ project(rct_common VERSION 0.1.0)
 
 include(cmake/rct_macros.cmake)
 
+find_package(Eigen3 REQUIRED)
+if(NOT EIGEN3_INCLUDE_DIRS)
+  set(EIGEN3_INCLUDE_DIRS ${EIGEN3_INCLUDE_DIR})
+endif()
+if(NOT TARGET Eigen3::Eigen)
+    find_package(Threads REQUIRED)
+    add_library(Eigen3::Eigen IMPORTED INTERFACE)
+    set_property(TARGET Eigen3::Eigen PROPERTY INTERFACE_COMPILE_DEFINITIONS ${EIGEN3_DEFINITIONS})
+    set_property(TARGET Eigen3::Eigen PROPERTY INTERFACE_INCLUDE_DIRECTORIES ${EIGEN3_INCLUDE_DIRS})
+endif()
+
 if(RCT_BUILD_TESTS)
   find_package(GTest QUIET)
   if ( NOT GTest_FOUND )
@@ -45,8 +56,17 @@ if(RCT_BUILD_TESTS)
 endif()
 
 add_library(${PROJECT_NAME} INTERFACE)
+target_compile_options(${PROJECT_NAME} INTERFACE -std=c++11)
+target_include_directories(${PROJECT_NAME} INTERFACE
+  "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
+  "$<INSTALL_INTERFACE:include>")
+target_link_libraries(${PROJECT_NAME} INTERFACE Eigen3::Eigen)
 
 rct_configure_package(${PROJECT_NAME})
+
+install(DIRECTORY include/${PROJECT_NAME}
+  DESTINATION include
+)
 
 install(FILES
   "${CMAKE_CURRENT_LIST_DIR}/cmake/rct_macros.cmake"
