+++ rct_common/cmake/rct_common-config.cmake.in
@@ -3,6 +3,9 @@
 set(@PROJECT_NAME@_FOUND ON)
 set_and_check(@PROJECT_NAME@_LIBRARY_DIRS "${PACKAGE_PREFIX_DIR}/lib")
 
+include(CMakeFindDependencyMacro)
+find_dependency(Eigen3)
+
 include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/rct_macros.cmake")
