+++ rct_common/package.xml
@@ -6,6 +6,7 @@
   <maintainer email="joseph.schornak@swri.org">Joseph Schornak</maintainer>
   <author email="joseph.schornak@swri.org">Joseph Schornak</author>
   <license>Apache 2.0</license>
+  <depend>eigen</depend>
   <export>
     <build_type>cmake</build_type>
   </export>
