+++ rct_examples/src/tools/solve_pnp.cpp
@@ -5,9 +5,9 @@
  */
 #include <rct_image_tools/image_utils.h>
 #include <rct_optimizations/pnp.h>
+#include <rct_common/print_utils.h>
 #include <rct_ros_tools/parameter_loaders.h>
 #include <rct_ros_tools/target_finder_plugin.h>
-#include <rct_ros_tools/print_utils.h>
 #include <rct_ros_tools/loader_utils.h>
 
 #include <opencv2/highgui.hpp>
@@ -18,6 +18,7 @@
 using namespace rct_optimizations;
 using namespace rct_image_tools;
 using namespace rct_ros_tools;
+using namespace rct_common;
 
 std::string WINDOW = "window";
