+++ rct_examples/src/tools/static_camera_extrinsic.cpp
@@ -1,8 +1,8 @@
 // Utilities for loading data sets and calib parameters from YAML files via ROS
+#include <rct_common/print_utils.h>
 #include <rct_ros_tools/data_set.h>
 #include <rct_ros_tools/parameter_loaders.h>
 #include <rct_ros_tools/target_finder_plugin.h>
-#include <rct_ros_tools/print_utils.h>
 #include <rct_ros_tools/loader_utils.h>
 // The calibration function for 'static camera' on robot wrist
 #include <rct_optimizations/extrinsic_hand_eye.h>
@@ -18,6 +18,7 @@
 using namespace rct_optimizations;
 using namespace rct_image_tools;
 using namespace rct_ros_tools;
+using namespace rct_common;
 
 std::string WINDOW = "window";
