+++ rct_image_tools/cmake/rct_image_tools-config.cmake.in
@@ -5,8 +5,11 @@ set_and_check(@PROJECT_NAME@_INCLUDE_DIRS "${PACKAGE_PREFIX_DIR}/include")
 set_and_check(@PROJECT_NAME@_LIBRARY_DIRS "${PACKAGE_PREFIX_DIR}/lib")
 
 include(CMakeFindDependencyMacro)
+find_dependency(Boost)
 find_dependency(Eigen3)
 find_dependency(OpenCV)
 find_dependency(yaml-cpp)
+find_dependency(rct_common)
+find_dependency(rct_optimizations)
 
 include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake")
